Disallow quotation marks (') from identifiers
Savile Row does not support ' characters in identifiers.
We could implement a workaround where we insert a marker (like
__PRIME__) for each ' character and recover these after a solution is
found.
But this will be too hairy, instead we will reject such identifiers for
now.
If somebody really needs to use a ' character as part of an identifier,
we can revisit this decision.