Travis times out if we ask it to both build Minion and run Conjure's (quick) tests within 50 minutes.
Solution: Do not run the tests if the Minion binary was built within this build. Do run them if the Minion binary was already in cache.
To check this we use `find -mmin -5`, i.e. if the minion executable file is younger then 5 minutes we skip the testing.

It's a good idea to check travis build logs for when this happens, since rerunning the build will likely use Minion from cache and run the tests.