Combining domains into constrained domains
Combining two domains into more constrained domains, by using min
and max expressions when both bounds are set, or by taking a bound
when one is set but not the other. Single domain ranges ignore the
other domain, even if it is another single value. Non-overlapping
domains are not found and should be an error.