Combining and ignoring in OpMin/OpMax
Combining values in OpMin and OpMax in the same matrix, instead of
creating deeply (in fact infinitely due to repeated
transformations) nested min and max expressions.

If an expression being added to the min or the max expression is
already contained within it, then it is not added, as there is no
need to compare it to itself.