Partially generalised variable size lifting
The rule for lifting a constraint on the size of a variable to an
attribute of the variable's domain has been partially generalised
to handle a single level of set/mset nested in a function on the
left hand side of the operator, and a function result on the right
hand side of it. It still supports direct references to sets and
msets in the same (messy) code.