Begun refactoring for better structure
I have begun a refactoring of the StrengthenVariables module to
have much better structure than what was there from the beginning.
The current structure will be a mix of the old one and of the one
that Ozgür used in the original branch, which was much simpler than
the one that was first used here.