Refactored structure in preparation for nesting
Restructured the code in preparation for nested domains. It is
better to do it before it is too late. If a rule does not fire for
a given constraint but the find variable's domain contains an inner
domain and the constraint is a forAll, then it will step inside the
domain and the constraint and try to apply the rule there, stepping
back up to fold the changes into the model at the correct
locations.