Using zippers of Expression by default
Using Zipper Expression Expression for constraints to add and to
remove, as well as constraints passed to rules. This makes climbing
back out of nested constraints easy and allows clever matching
rules for constraints nested in forAlls without any need of a
clever way of getting back out of the forAll.