Renamed to ModelStrengthening and logging rules
Renamed variable strengthening to model strengthening, which is
more general.

Added logging of successful rules (with user-specified level and
only if the flag is set). Each rule is associated with a string
and if it fires, the string is printed, along with the variable
that was modified (in its unmodified form) or the constraints that
were added.