fix(docker): install conjure to /opt/conjure not ~/.local/bin
In the container, install Conjure and its dependencies to /opt/conjure
instead of ~/.local/bin.

This allows the user inside the container to be changed and the binaries
still be accessible.

I encountered this issue when using distrobox (https://distrobox.it/)
with the Conjure container. This setups a container that has access to
your /home folder, and runs under your user.

More generally, it seems common (in my very limited experience ) to use
non-root users inside docker containers for security reasons. For
example, if you mount a folder in the container, and the inside of the
container runs with root permissions, then it also has root permissions
for that mounted volume.

I do not know enough about Docker to say whether this is a security
issue in this case, so this commit retains the current behaviour of root
being the default user.